creating randomizer that adds characters, moved Generators to its own file
