+++ GitHubWebUserFinder.Tests/GitHubWebUserFinder.Tests.csproj
@@ -103,6 +103,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Controllers\ResultControllerTest.cs" />
+    <Compile Include="Models\Generators.cs" />
     <Compile Include="Models\GitHubUserTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Controllers\HomeControllerTest.cs" />
